am243x: threadx: Remove threadx component.
- removed eclipse_threadx directory.
- moved threadx and filex directories under kernel and fs components respectively.

Fixes: SITSW-5614
