+++ .project/device/project_am243x.js
@@ -4,14 +4,14 @@ const component_file_list = [
     "source/board/.project/project.js",
     "source/cmsis/.project/project.js",
     "source/drivers/.project/project.js",
-    "source/eclipse_threadx/threadx/.project/project.js",
-    "source/eclipse_threadx/filex/.project/project.js",
+    "source/fs/filex/.project/project.js",
     "source/fs/freertos_fat/.project/project.js",
     "source/fs/littlefs/.project/project.js",
     "source/middleware/.project/project.js",
     "source/sdl/.project/project.js",
     "source/kernel/nortos/.project/project.js",
     "source/kernel/freertos/.project/project.js",
+    "source/kernel/threadx/.project/project.js",
     "source/mathlib/.project/project.js",
     "source/pru_io/.project/project.js",
     "source/security/.project/project.js",
@@ -164,11 +164,9 @@ const example_file_list = [
     "examples/drivers/udma/udma_memcpy_polling/.project/project.js",
     "examples/drivers/udma/udma_sw_trigger/.project/project.js",
     "examples/drivers/watchdog/watchdog_interrupt/.project/project.js",
-    "examples/eclipse_threadx/threadx/hello_world/.project/project.js",
-    "examples/eclipse_threadx/threadx/task_switch/.project/project.js",
-    "examples/eclipse_threadx/filex/hello_world/.project/project.js",
     "examples/empty/.project/project_freertos.js",
     "examples/empty/.project/project_nortos.js",
+    "examples/fs/filex/hello_world/.project/project.js",
     "examples/hello_world/.project/project.js",
     "examples/hello_world_cpp/.project/project.js",
     "examples/kernel/dpl/dpl_demo/.project/project.js",
@@ -178,6 +176,8 @@ const example_file_list = [
     "examples/kernel/freertos/task_switch/.project/project.js",
     "examples/kernel/nortos/wfi_standby_demo/.project/project_am243x.js",
     "examples/kernel/nortos/basic_smart_placement/.project/project_nortos.js",
+    "examples/kernel/threadx/hello_world/.project/project.js",
+    "examples/kernel/threadx/task_switch/.project/project.js",
     "examples/mathlib/benchmark/.project/project.js",
     "examples/security/crypto/sa2ul_aes/crypto_aes_cbc_256/.project/project.js",
     "examples/security/crypto/sa2ul_aes/crypto_aes_cbc_128/.project/project.js",
@@ -314,8 +314,6 @@ const example_file_list = [
     "test/drivers/soc/soc_m4f/.project/project.js",
     "test/drivers/uart/.project/project.js",
     "test/drivers/gp_timer/.project/project.js",
-    "test/eclipse_threadx/threadx/.project/project.js",
-    "test/eclipse_threadx/filex/.project/project.js",
     "test/kernel/dpl/.project/project.js",
     "test/kernel/dpl/dpl_ut/nortos/common/ClockP_nortos/.project/project.js",
     "test/kernel/dpl/dpl_ut/nortos/common/HeapP_nortos/.project/project.js",
