+++ .project/templates/product.json.xdt
@@ -25,7 +25,6 @@
         "/kernel/dpl",
         "/drivers/drivers",
         "/board/board",
-        "/eclipse_threadx/eclipse_threadx",
         "/fs/fs",
         "/networking/networking",
         "/security/security",
