+++ examples/fs/filex/hello_world/am243x-lp/r5fss0-0_threadx/example.syscfg
@@ -8,9 +8,8 @@
 /**
  * Import the modules used in this configuration.
  */
-const filex           = scripting.addModule("/eclipse_threadx/filex/filex", {}, false);
+const filex           = scripting.addModule("/fs/filex/filex", {}, false);
 const filex1          = filex.addInstance();
-const threadx         = scripting.addModule("/eclipse_threadx/threadx/threadx");
 const debug_log       = scripting.addModule("/kernel/dpl/debug_log");
 const mpu_armv7       = scripting.addModule("/kernel/dpl/mpu_armv7", {}, false);
 const mpu_armv71      = mpu_armv7.addInstance();
