+++ examples/kernel/threadx/hello_world/.project/project.js
