+++ examples/kernel/threadx/hello_world/.project/project_am243x.js
