+++ examples/kernel/threadx/hello_world/am243x-evm/r5fss0-0_threadx/example.syscfg
