+++ examples/kernel/threadx/hello_world/threadx_hello_world.c
