+++ examples/kernel/threadx/task_switch/.project/project.js
