+++ examples/kernel/threadx/task_switch/.project/project_am243x.js
