+++ examples/kernel/threadx/task_switch/am243x-evm/r5fss0-0_threadx/example.syscfg
