+++ examples/kernel/threadx/task_switch/am243x-lp/r5fss0-0_threadx/example.syscfg
