+++ examples/kernel/threadx/task_switch/threadx_task_switch.c
