+++ source/kernel/threadx/dpl/common/DebugP_shmLogReader_threadx.c
