+++ source/kernel/threadx/dpl/common/SemaphoreP_threadx.c
