+++ source/kernel/threadx/dpl/r5/ClockP_threadx_r5.c
@@ -30,7 +30,7 @@
  *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <eclipse_threadx/threadx/dpl/common/ClockP_threadx_priv.h>
+#include <kernel/threadx/dpl/common/ClockP_threadx_priv.h>
 #include <kernel/dpl/TimerP.h>
 
 void ClockP_timerClearOverflowInt(uint32_t timerBaseAddr)
