+++ source/kernel/threadx/ports/ti_arm_gcc_clang_cortex_r5/src/tx_thread_context_restore.S
