+++ test/fs/filex/.project/project.js
