+++ test/fs/filex/.project/project_am243x.js
@@ -4,7 +4,7 @@ let device = "am243x";
 
 const files = {
     common: [
-        "filex_hello_world.c",
+        "test_filex.c",
         "main.c",
     ],
 };
@@ -15,25 +15,29 @@ const files = {
 const filedirs = {
     common: [
         "..",       /* core_os_combo base */
+        "../..",    /* Board base */
         "../../..", /* Example base */
     ],
 };
 
 const libdirs = {
     common: [
-        "${MCU_PLUS_SDK_PATH}/source/eclipse_threadx/filex/lib",
-        "${MCU_PLUS_SDK_PATH}/source/eclipse_threadx/threadx/lib",
         "${MCU_PLUS_SDK_PATH}/source/drivers/lib",
+        "${MCU_PLUS_SDK_PATH}/source/fs/filex/lib",
+        "${MCU_PLUS_SDK_PATH}/source/kernel/threadx/lib",
         "${MCU_PLUS_SDK_PATH}/source/board/lib",
+        "${MCU_PLUS_SDK_PATH}/test/unity/lib",
     ],
 };
 
 const includes = {
     common: [
-        "${MCU_PLUS_SDK_PATH}/source/eclipse_threadx/threadx/threadx_src/common/inc",
-        "${MCU_PLUS_SDK_PATH}/source/eclipse_threadx/threadx/ports/ti_arm_gcc_clang_cortex_r5/inc",
-        "${MCU_PLUS_SDK_PATH}/source/eclipse_threadx/filex/filex_src/common/inc",
-        "${MCU_PLUS_SDK_PATH}/source/eclipse_threadx/filex/filex_src/ports/generic/inc",
+        "${MCU_PLUS_SDK_PATH}/test/unity/",
+        "${MCU_PLUS_SDK_PATH}/source/kernel/threadx/threadx_src/common/inc",
+        "${MCU_PLUS_SDK_PATH}/source/kernel/threadx/ports/ti_arm_gcc_clang_cortex_r5/inc",
+        "${MCU_PLUS_SDK_PATH}/source/fs/filex/filex_src/common/inc",
+        "${MCU_PLUS_SDK_PATH}/source/fs/filex/filex_mmcsd",
+        "${MCU_PLUS_SDK_PATH}/source/fs/filex/filex_src/ports/generic/inc",
     ],
 };
 
@@ -43,17 +47,10 @@ const libs = {
         "filex.am243x.r5f.ti-arm-clang.${ConfigName}.lib",
         "drivers.am243x.r5f.ti-arm-clang.${ConfigName}.lib",
         "board.am243x.r5f.ti-arm-clang.${ConfigName}.lib",
+        "unity.am243x.r5f.ti-arm-clang.${ConfigName}.lib",
     ],
 };
 
-const libs_gcc = {
-    common: [
-        "threadx.am243x.r5f.gcc-armv7.${ConfigName}.lib",
-        "filex.am243x.r5f.gcc-armv7.${ConfigName}.lib",
-        "drivers.am243x.r5f.gcc-armv7.${ConfigName}.lib",
-        "board.am243x.r5f.gcc-armv7.${ConfigName}.lib",
-    ],
-};
 
 const lnkfiles = {
     common: [
@@ -63,30 +60,13 @@ const lnkfiles = {
 
 const syscfgfile = "../example.syscfg";
 
-const readmeDoxygenPageTag = "EXAMPLES_ECLIPSE_THREADX_FILEX_HELLO_WORLD";
-
-const templates_gcc =
-[
-    {
-        input: ".project/templates/am243x/common/linker_r5f_gcc.cmd.xdt",
-        output: "linker.cmd",
-    },
-    {
-        input: ".project/templates/am243x/threadx/main_threadx.c.xdt",
-        output: "../main.c",
-        options: {
-            entryFunction: "filex_hello_world_main",
-        },
-    }
-];
-
 const templates =
 [
     {
         input: ".project/templates/am243x/threadx/main_threadx.c.xdt",
         output: "../main.c",
         options: {
-            entryFunction: "filex_hello_world_main",
+            entryFunction: "test_main",
         },
     }
 ];
@@ -95,7 +75,7 @@ const buildOptionCombos = [
     { device: device, cpu: "r5fss0-0", cgt: "ti-arm-clang", board: "am243x-evm", os: "threadx"},
     { device: device, cpu: "r5fss0-0", cgt: "gcc-armv7",    board: "am243x-evm", os: "threadx"},
     { device: device, cpu: "r5fss0-0", cgt: "ti-arm-clang", board: "am243x-lp", os: "threadx"},
-    { device: device, cpu: "r5fss0-0", cgt: "gcc-armv7",    board: "am243x-lp", os: "threadx"}
+    { device: device, cpu: "r5fss0-0", cgt: "gcc-armv7",    board: "am243x-lp", os: "threadx"},
 ];
 
 function getComponentProperty() {
@@ -103,10 +83,9 @@ function getComponentProperty() {
 
     property.dirPath = path.resolve(__dirname, "..");
     property.type = "executable";
-    property.name = "filex_hello_world";
-    property.isInternal = false;
-    property.tirexResourceSubClass = [ "example.gettingstarted" ];
-    property.description = "A simple \"Hello, World\" example for the FileX file system. "
+    property.name = "test_filex";
+    property.isInternal = true;
+    property.skipProjectSpec = true;
     property.buildOptionCombos = buildOptionCombos;
 
     return property;
@@ -117,17 +96,14 @@ function getComponentBuildProperty(buildOption) {
 
     build_property.files = files;
     build_property.filedirs = filedirs;
+    build_property.includes = includes;
     build_property.libdirs = libdirs;
     build_property.lnkfiles = lnkfiles;
     build_property.syscfgfile = syscfgfile;
-    build_property.readmeDoxygenPageTag = readmeDoxygenPageTag;
-    build_property.includes = includes;
-    build_property.libdirs = libdirs;
-    build_property.templates = templates;
-    if(buildOption.cpu.match(/gcc*/)) {
-        build_property.libs = libs_gcc;
-    } else {
+
+    if(buildOption.cpu.match(/r5f*/)) {
         build_property.libs = libs;
+        build_property.templates = templates;
     }
 
     return build_property;
