+++ test/fs/filex/am243x-evm/r5fss0-0_threadx/example.syscfg
