+++ test/kernel/threadx/am243x-evm/r5fss0-0_threadx/example.syscfg
