+++ test/kernel/threadx/am243x-lp/r5fss0-0_threadx/example.syscfg
